pyproject.toml
[tool.poetry]
name = "funny-tweets-re"
version = "0.1.0"
description = "Funny tweets retweet bot"
authors = ["Claudio <claudio.work69@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests-oauthlib = "^1.3.1"
firebase-admin = "^5.2.0"
pendulum = "^2.1.2"
schedule = "^1.1.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
beeprint = "^2.4.10"
flake8-black = "^0.3.3"
pre-commit = "^2.20.0"
requests-mock = "^1.9.3"
pytest = "^7.1.2"
pytest-mock = "^3.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
